# dkim: Only compile regex once