@56KBs 2023-12-15
I've been reading through the codebase some more and have spotted some other areas of code that could be modified to reduce CPU time/allocations/memory use - Would you be interested in me raising PRs for these @emersion? If so would you prefer one PR with multiple enhancements committed or one PR per enhancement?