@emersion 2023-12-15
Feel free to send PR(s). Note that I would rather keep the codebase simple: as long as the optimized versions don't decrease readability too much I'd welcome such changes. I don't mind as long as the changes are [properly split logically per commit](https://www.bitsnbites.eu/git-history-work-log-vs-recipe/), so as you prefer: either one PR and you craft the logical commits, either multiple PRs and I can squash when merging.